Consolidate create dialogs and queue-item form
- DEFAULT_GRANTS + CreateAccessSection replace the grant literal and
  the copy-pasted advanced-access Accordion/ShareControl block in the
  five create dialogs (document keeps its collapsed-by-default state
  via defaultOpen).
- CreateQueueDialog and CreateCounterGroupDialog were line-for-line
  twins; both are now thin wrappers over one CreateToolDialog taking a
  five-field concrete config (namespace, title/description keys, id
  prefix, create hook). Call sites unchanged.
- useQueueItemForm extracts the ~90%-shared field state and picker
  option lists of Add/EditQueueItemDialog, preserving the server
  typeaheads (enabled gating, page-size limits) from the earlier
  picker work. Submit payloads and edit-only state stay per-dialog.

Refs #870
